Extracted GridPos
